pipelines/batch/requirements.txt
boto3==1.34.25
pandas==2.2.0
pyarrow==15.0.0
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
